+++ Riot/Categories/MXRoom+Riot.m
@@ -20,7 +20,7 @@
 
 #import "AvatarGenerator.h"
 #import "MatrixKit.h"
-
+#import "GeneratedInterface-Swift.h"
 #import <objc/runtime.h>
 
 @implementation MXRoom (Riot)
@@ -80,8 +80,15 @@
     // Check whether an override rule has been defined with the roomm id as rule id.
     // This kind of rule is created to mute the room
     MXPushRule* rule = [self getOverrideRoomPushRule];
-    if (rule)
+    if (rule && rule.enabled)
     {
+        // Support for MSC3987: The dont_notify push rule action is deprecated.
+        if (rule.actions.count == 0)
+        {
+            return true;
+        }
+        
+        // Support deprecated dont_notify push rule action for compatibility purposes.
         for (MXPushRuleAction *ruleAction in rule.actions)
         {
             if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
@@ -98,7 +105,7 @@
                         
                         if (key && pattern && [key isEqualToString:@"room_id"] && [pattern isEqualToString:self.roomId])
                         {
-                            return rule.enabled;
+                            return true;
                         }
                     }
                 }
@@ -113,13 +120,20 @@
 {
     // Check push rules at room level
     MXPushRule *rule = [self getRoomPushRule];
-    if (rule)
+    if (rule && rule.enabled)
     {
+        // Support for MSC3987: The dont_notify push rule action is deprecated.
+        if (rule.actions.count == 0)
+        {
+            return true;
+        }
+        
+        // Support deprecated dont_notify push rule action for compatibility purposes.
         for (MXPushRuleAction *ruleAction in rule.actions)
         {
             if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
             {
-                return rule.enabled;
+                return true;
             }
         }
     }
@@ -178,12 +192,21 @@
         // check if the user did not define one
         BOOL hasDontNotifyRule = NO;
         
-        for (MXPushRuleAction *ruleAction in rule.actions)
+        // Support for MSC3987: The dont_notify push rule action is deprecated.
+        if (rule.actions.count == 0)
         {
-            if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
+            hasDontNotifyRule = YES;
+        }
+        else
+        {
+            // Support deprecated dont_notify push rule action for compatibility purposes.
+            for (MXPushRuleAction *ruleAction in rule.actions)
             {
-                hasDontNotifyRule = YES;
-                break;
+                if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
+                {
+                    hasDontNotifyRule = YES;
+                    break;
+                }
             }
         }
         
@@ -256,12 +279,21 @@
         // check if the user did not define one
         BOOL hasDontNotifyRule = NO;
         
-        for (MXPushRuleAction *ruleAction in rule.actions)
+        // Support for MSC3987: The dont_notify push rule action is deprecated.
+        if (rule.actions.count == 0)
         {
-            if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
+            hasDontNotifyRule = YES;
+        }
+        else
+        {
+            // Support deprecated dont_notify push rule action for compatibility purposes.
+            for (MXPushRuleAction *ruleAction in rule.actions)
             {
-                hasDontNotifyRule = YES;
-                break;
+                if (ruleAction.actionType == MXPushRuleActionTypeDontNotify)
+                {
+                    hasDontNotifyRule = YES;
+                    break;
+                }
             }
         }
         
@@ -331,30 +363,10 @@
     {
         [self.mxSession.crypto trustLevelSummaryForUserIds:@[userId] forceDownload:NO success:^(MXUsersTrustLevelSummary *usersTrustLevelSummary) {
             
-            UserEncryptionTrustLevel userEncryptionTrustLevel;
-            double trustedDevicesPercentage = usersTrustLevelSummary.trustedDevicesProgress.fractionCompleted;
-            
-            if (trustedDevicesPercentage >= 1.0)
-            {
-                userEncryptionTrustLevel = UserEncryptionTrustLevelTrusted;
-            }
-            else if (trustedDevicesPercentage == 0.0)
-            {
-                // Verify if the user has the user has cross-signing enabled
-                if ([self.mxSession.crypto.crossSigning crossSigningKeysForUser:userId])
-                {
-                    userEncryptionTrustLevel = UserEncryptionTrustLevelNotVerified;
-                }
-                else
-                {
-                    userEncryptionTrustLevel = UserEncryptionTrustLevelNoCrossSigning;
-                }
-            }
-            else
-            {
-                userEncryptionTrustLevel = UserEncryptionTrustLevelWarning;
-            }
-            
+            MXCrossSigningInfo *crossSigningInfo = [self.mxSession.crypto.crossSigning crossSigningKeysForUser:userId];
+            EncryptionTrustLevel *encryption = [[EncryptionTrustLevel alloc] init];
+            UserEncryptionTrustLevel userEncryptionTrustLevel = [encryption userTrustLevelWithCrossSigning:crossSigningInfo
+                                                                                    trustedDevicesProgress:usersTrustLevelSummary.trustedDevicesProgress];
             onComplete(userEncryptionTrustLevel);
             
         } failure:^(NSError *error) {