+++ Riot/Modules/Integrations/Widgets/Jitsi/JitsiViewController.m
@@ -35,6 +35,7 @@ static NSString * _Nonnull kRCTTextViewClassName = @"RCTTextView";
  Some feature flags defined in https://github.com/jitsi/jitsi-meet/blob/master/react/features/base/flags/constants.js
  */
 static NSString * _Nonnull kJitsiFeatureFlagChatEnabled = @"chat.enabled";
+static NSString * _Nonnull kJitsiFeatureFlagScreenSharingEnabled = @"ios.screensharing.enabled";
 
 // TODO: Tchap: VoIP support (Jitsi)
 @interface JitsiViewController () </*PictureInPicturable,*/ JitsiMeetViewDelegate>
@@ -279,6 +280,7 @@ static NSString * _Nonnull kJitsiFeatureFlagChatEnabled = @"chat.enabled";
                                                                     andAvatar:avatarUrl];
             builder.token = self.jwtToken;
             [builder setFeatureFlag:kJitsiFeatureFlagChatEnabled withBoolean:NO];
+            [builder setFeatureFlag:kJitsiFeatureFlagScreenSharingEnabled withBoolean: YES];
         }];
         
         [self.jitsiMeetView join:jitsiMeetConferenceOptions];