+++ Riot/Modules/Room/RoomViewController.m
@@ -96,7 +96,7 @@ static CGSize kThreadListBarButtonItemImageSize;
 @interface RoomViewController () <UISearchBarDelegate, UIGestureRecognizerDelegate, UIScrollViewAccessibilityDelegate, RoomTitleViewTapGestureDelegate, RoomParticipantsViewControllerDelegate, MXKRoomMemberDetailsViewControllerDelegate, MXServerNoticesDelegate, RoomContextualMenuViewControllerDelegate,
     ReactionsMenuViewModelCoordinatorDelegate, EditHistoryCoordinatorBridgePresenterDelegate, MXKDocumentPickerPresenterDelegate, EmojiPickerCoordinatorBridgePresenterDelegate,
     ReactionHistoryCoordinatorBridgePresenterDelegate, CameraPresenterDelegate, MediaPickerCoordinatorBridgePresenterDelegate,
-    RoomDataSourceDelegate/*, RoomCreationModalCoordinatorBridgePresenterDelegate*/, RoomInfoCoordinatorBridgePresenterDelegate/*, DialpadViewControllerDelegate, RemoveJitsiWidgetViewDelegate*/, VoiceMessageControllerDelegate /*, SpaceDetailPresenterDelegate*/, UserSuggestionCoordinatorBridgeDelegate/*, ThreadsCoordinatorBridgePresenterDelegate, ThreadsBetaCoordinatorBridgePresenterDelegate, MXThreadingServiceDelegate, RoomParticipantsInviteCoordinatorBridgePresenterDelegate*/, RoomInputToolbarViewDelegate/*, ComposerCreateActionListBridgePresenterDelegate*/>
+    RoomDataSourceDelegate/*, RoomCreationModalCoordinatorBridgePresenterDelegate*/, RoomInfoCoordinatorBridgePresenterDelegate/*, DialpadViewControllerDelegate, RemoveJitsiWidgetViewDelegate*/, VoiceMessageControllerDelegate /*, SpaceDetailPresenterDelegate*/, CompletionSuggestionCoordinatorBridgeDelegate/*, ThreadsCoordinatorBridgePresenterDelegate, ThreadsBetaCoordinatorBridgePresenterDelegate, MXThreadingServiceDelegate, RoomParticipantsInviteCoordinatorBridgePresenterDelegate*/, RoomInputToolbarViewDelegate/*, ComposerCreateActionListBridgePresenterDelegate*/>
 {
     
     // The preview header
@@ -221,8 +221,8 @@ static CGSize kThreadListBarButtonItemImageSize;
 @property (nonatomic, strong) ShareManager *shareManager;
 @property (nonatomic, strong) EventMenuBuilder *eventMenuBuilder;
 
-@property (nonatomic, strong) UserSuggestionCoordinatorBridge *userSuggestionCoordinator;
-@property (nonatomic, weak) IBOutlet UIView *userSuggestionContainerView;
+@property (nonatomic, strong) CompletionSuggestionCoordinatorBridge *completionSuggestionCoordinator;
+@property (nonatomic, weak) IBOutlet UIView *completionSuggestionContainerView;
 
 @property (nonatomic, readwrite) RoomDisplayConfiguration *displayConfiguration;
 
@@ -414,7 +414,7 @@ static CGSize kThreadListBarButtonItemImageSize;
     
     [self setupActions];
     
-    [self setupUserSuggestionViewIfNeeded];
+    [self setupCompletionSuggestionViewIfNeeded];
     
     [self.topBannersStackView vc_removeAllSubviews];
 }
@@ -691,7 +691,7 @@ static CGSize kThreadListBarButtonItemImageSize;
     {
         // Retrieve the potential message partially typed during last room display.
         // Note: We have to wait for viewDidAppear before updating growingTextView (viewWillAppear is too early)
-        self.inputToolbarView.attributedTextMessage = self.roomDataSource.partialAttributedTextMessage;
+        [self.inputToolbarView setPartialContent:self.roomDataSource.partialAttributedTextMessage];
     }
     
     [self setMaximisedToolbarIsHiddenIfNeeded: NO];
@@ -1085,13 +1085,15 @@ static CGSize kThreadListBarButtonItemImageSize;
     
     [VoiceMessageMediaServiceProvider.sharedProvider setCurrentRoomSummary:dataSource.room.summary];
     _voiceMessageController.roomId = dataSource.roomId;
-
-    _userSuggestionCoordinator = [[UserSuggestionCoordinatorBridge alloc] initWithMediaManager:self.roomDataSource.mxSession.mediaManager
+    
+    _completionSuggestionCoordinator = [[CompletionSuggestionCoordinatorBridge alloc] initWithMediaManager:self.roomDataSource.mxSession.mediaManager
                                                                                           room:dataSource.room
                                                                                         userID:self.roomDataSource.mxSession.myUserId];
-    _userSuggestionCoordinator.delegate = self;
+    _completionSuggestionCoordinator.delegate = self;
     
-    [self setupUserSuggestionViewIfNeeded];
+    [self setupCompletionSuggestionViewIfNeeded];
+
+    [self updateRoomInputToolbarViewClassIfNeeded];
     
     [self updateTopBanners];
 }
@@ -1192,6 +1194,12 @@ static CGSize kThreadListBarButtonItemImageSize;
 - (void)updateRoomInputToolbarViewClassIfNeeded
 {
     Class roomInputToolbarViewClass = [RoomViewController mainToolbarClass];
+
+    // If RTE is enabled, delay the toolbar setup until `completionSuggestionCoordinator` is ready.
+    if (roomInputToolbarViewClass == WysiwygInputToolbarView.class && _completionSuggestionCoordinator == nil)
+    {
+        return;
+    }
     
     BOOL shouldDismissContextualMenu = NO;
     
@@ -1281,6 +1289,8 @@ static CGSize kThreadListBarButtonItemImageSize;
 - (BOOL)sendAsIRCStyleCommandIfPossible:(NSString*)string
 {
     // Override the default behavior for `/join` command in order to open automatically the joined room
+
+    NSString* kMXKSlashCmdJoinRoom = [MXKSlashCommandsHelper commandNameFor:MXKSlashCommandJoinRoom];
     
     if ([string hasPrefix:kMXKSlashCmdJoinRoom])
     {
@@ -1317,7 +1327,7 @@ static CGSize kThreadListBarButtonItemImageSize;
         else
         {
             // Display cmd usage in text input as placeholder
-            self.inputToolbarView.placeholder = @"Usage: /join <room_alias>";
+            self.inputToolbarView.placeholder = [MXKSlashCommandsHelper commandUsageFor:MXKSlashCommandJoinRoom];
         }
         return YES;
     }
@@ -2055,7 +2065,7 @@ static CGSize kThreadListBarButtonItemImageSize;
         {
             [userPictureView vc_setRoomAvatarImageWith:self.directChatTargetUser.avatarUrl
                                                 roomId:self.directChatTargetUser.userId
-                                           displayName:self.directChatTargetUser.displayname
+                                           displayName:self.directChatTargetUser.displayname ?: self.directChatTargetUser.userId
                                           mediaManager:self.mainSession.mediaManager];
         }
     }
@@ -2738,14 +2748,14 @@ static CGSize kThreadListBarButtonItemImageSize;
     }
 }
 
-- (void)setupUserSuggestionViewIfNeeded
+- (void)setupCompletionSuggestionViewIfNeeded
 {
     if(!self.isViewLoaded) {
         return;
     }
     
-    UIViewController *suggestionsViewController = self.userSuggestionCoordinator.toPresentable;
-
+    UIViewController *suggestionsViewController = self.completionSuggestionCoordinator.toPresentable;
+    
     if (!suggestionsViewController)
     {
         return;
@@ -2754,13 +2764,13 @@ static CGSize kThreadListBarButtonItemImageSize;
     [suggestionsViewController.view setTranslatesAutoresizingMaskIntoConstraints:NO];
 
     [self addChildViewController:suggestionsViewController];
-    [self.userSuggestionContainerView addSubview:suggestionsViewController.view];
-
-    [NSLayoutConstraint activateConstraints:@[[suggestionsViewController.view.topAnchor constraintEqualToAnchor:self.userSuggestionContainerView.topAnchor],
-                                              [suggestionsViewController.view.leadingAnchor constraintEqualToAnchor:self.userSuggestionContainerView.leadingAnchor],
-                                              [suggestionsViewController.view.trailingAnchor constraintEqualToAnchor:self.userSuggestionContainerView.trailingAnchor],
-                                              [suggestionsViewController.view.bottomAnchor constraintEqualToAnchor:self.userSuggestionContainerView.bottomAnchor],]];
-
+    [self.completionSuggestionContainerView addSubview:suggestionsViewController.view];
+    
+    [NSLayoutConstraint activateConstraints:@[[suggestionsViewController.view.topAnchor constraintEqualToAnchor:self.completionSuggestionContainerView.topAnchor],
+                                              [suggestionsViewController.view.leadingAnchor constraintEqualToAnchor:self.completionSuggestionContainerView.leadingAnchor],
+                                              [suggestionsViewController.view.trailingAnchor constraintEqualToAnchor:self.completionSuggestionContainerView.trailingAnchor],
+                                              [suggestionsViewController.view.bottomAnchor constraintEqualToAnchor:self.completionSuggestionContainerView.bottomAnchor],]];
+    
     [suggestionsViewController didMoveToParentViewController:self];
 }
 
@@ -5176,22 +5186,22 @@ static CGSize kThreadListBarButtonItemImageSize;
  
 - (void)roomInputToolbarViewDidChangeTextMessage:(RoomInputToolbarView *)toolbarView
 {
-    [self.userSuggestionCoordinator processTextMessage:toolbarView.textMessage];
+    [self.completionSuggestionCoordinator processTextMessage:toolbarView.textMessage];
 }
 
 - (void)didDetectTextPattern:(SuggestionPatternWrapper *)suggestionPattern
 {
-    [self.userSuggestionCoordinator processSuggestionPattern:suggestionPattern];
+    [self.completionSuggestionCoordinator processSuggestionPattern:suggestionPattern];
 }
 
-- (UserSuggestionViewModelContextWrapper *)userSuggestionContext
+- (CompletionSuggestionViewModelContextWrapper *)completionSuggestionContext
 {
-    return [self.userSuggestionCoordinator sharedContext];
+    return [self.completionSuggestionCoordinator sharedContext];
 }
 
 - (MXMediaManager *)mediaManager
 {
-    return self.roomDataSource.mxSession.mediaManager;
+    return self.mainSession.mediaManager;
 }
 
 - (void)roomInputToolbarViewDidOpenActionMenu:(RoomInputToolbarView*)toolbarView
@@ -5266,7 +5276,7 @@ static CGSize kThreadListBarButtonItemImageSize;
         if (readyToSend) {
             BOOL isMessageAHandledCommand = NO;
             // "/me" command is supported with Pills in RoomDataSource.
-            if (![attributedTextMessage.string hasPrefix:kMXKSlashCmdEmote])
+            if (![attributedTextMessage.string hasPrefix:[MXKSlashCommandsHelper commandNameFor:MXKSlashCommandEmote]])
             {
                 // Other commands currently work with identifiers (e.g. ban, invite, op, etc).
                 NSString *message;
@@ -5289,6 +5299,11 @@ static CGSize kThreadListBarButtonItemImageSize;
         }
         // Errors are handled at the request level. This should be improved in case of code rewriting.
     }];
+}
+
+- (void)roomInputToolbarView:(MXKRoomInputToolbarView *)toolbarView shouldStorePartialContent:(NSAttributedString *)partialAttributedTextMessage
+{
+    self.roomDataSource.partialAttributedTextMessage = partialAttributedTextMessage;
 }
 
 #pragma mark - MXKRoomMemberDetailsViewControllerDelegate
@@ -6133,7 +6148,7 @@ static CGSize kThreadListBarButtonItemImageSize;
             if (self.saveProgressTextInput)
             {
                 // Restore the potential message partially typed before jump to last unread messages.
-                self.inputToolbarView.attributedTextMessage = roomDataSource.partialAttributedTextMessage;
+                [self.inputToolbarView setPartialContent:roomDataSource.partialAttributedTextMessage];
             }
         };
 
@@ -6387,21 +6402,10 @@ static CGSize kThreadListBarButtonItemImageSize;
                 self->currentAlert = nil;
                 
                 // Acknowledge the existence of all devices
-                [self startActivityIndicator];
+                self->unknownDevices = nil;
                 
-                if (![self.mainSession.crypto isKindOfClass:[MXLegacyCrypto class]])
-                {
-                    MXLogFailure(@"[RoomVC] eventDidChangeSentState: Only legacy crypto supports manual setting of known devices");
-                    return;
-                }
-                [(MXLegacyCrypto *)self.mainSession.crypto setDevicesKnown:self->unknownDevices complete:^{
-                    
-                    self->unknownDevices = nil;
-                    [self stopActivityIndicator];
-                    
-                    // And resend pending messages
-                    [self resendAllUnsentMessages];
-                }];
+                // And resend pending messages
+                [self resendAllUnsentMessages];
             }
             
         }]];
@@ -7522,23 +7526,47 @@ static CGSize kThreadListBarButtonItemImageSize;
         return;
     }
     
+    NSMutableArray<NSIndexPath *> *rowsToReload = [[NSMutableArray alloc] init];
+    // Get the current hightlighted event because we will need to reload it
+    NSString *currentHiglightedEventId = self.customizedRoomDataSource.highlightedEventId;
+    if (currentHiglightedEventId)
+    {
+        NSInteger currentHiglightedRow = [self.roomDataSource indexOfCellDataWithEventId:currentHiglightedEventId];
+        if (currentHiglightedRow != NSNotFound)
+        {
+            NSIndexPath *indexPath = [NSIndexPath indexPathForRow:currentHiglightedRow inSection:0];
+            if ([[self.bubblesTableView indexPathsForVisibleRows] containsObject:indexPath])
+            {
+                [rowsToReload addObject:indexPath];
+            }
+        }
+    }
+    
     self.customizedRoomDataSource.highlightedEventId = eventId;
     
+    // Add the new highligted event to the list of rows to reload
     NSIndexPath *indexPath = [NSIndexPath indexPathForRow:row inSection:0];
-    if ([[self.bubblesTableView indexPathsForVisibleRows] containsObject:indexPath])
+    BOOL indexPathIsVisible = [[self.bubblesTableView indexPathsForVisibleRows] containsObject:indexPath];
+    if (indexPathIsVisible)
     {
-        [self.bubblesTableView reloadRowsAtIndexPaths:@[indexPath]
+        [rowsToReload addObject:indexPath];
+    }
+    
+    // Reload rows
+    if (rowsToReload.count > 0)
+    {
+        [self.bubblesTableView reloadRowsAtIndexPaths:rowsToReload
                                      withRowAnimation:UITableViewRowAnimationNone];
-        [self.bubblesTableView scrollToRowAtIndexPath:indexPath
-                                     atScrollPosition:UITableViewScrollPositionMiddle
-                                             animated:YES];
     }
-    else if ([self.bubblesTableView vc_hasIndexPath:indexPath])
+    
+    // Scroll to the newly highlighted row
+    if (indexPathIsVisible || [self.bubblesTableView vc_hasIndexPath:indexPath])
     {
         [self.bubblesTableView scrollToRowAtIndexPath:indexPath
                                      atScrollPosition:UITableViewScrollPositionMiddle
                                              animated:YES];
     }
+
     if (completion)
     {
         completion();
@@ -8109,9 +8137,9 @@ static CGSize kThreadListBarButtonItemImageSize;
 //    [[LegacyAppDelegate theDelegate] openSpaceWithId:spaceId];
 //}
 
-#pragma mark - UserSuggestionCoordinatorBridgeDelegate
+#pragma mark - CompletionSuggestionCoordinatorBridgeDelegate
 
-- (void)userSuggestionCoordinatorBridge:(UserSuggestionCoordinatorBridge *)coordinator
+- (void)completionSuggestionCoordinatorBridge:(CompletionSuggestionCoordinatorBridge *)coordinator
              didRequestMentionForMember:(MXRoomMember *)member
                             textTrigger:(NSString *)textTrigger
 {
@@ -8119,16 +8147,32 @@ static CGSize kThreadListBarButtonItemImageSize;
     [self mention:member];
 }
 
-- (void)userSuggestionCoordinatorBridgeDidRequestMentionForRoom:(UserSuggestionCoordinatorBridge *)coordinator
+- (void)completionSuggestionCoordinatorBridgeDidRequestMentionForRoom:(CompletionSuggestionCoordinatorBridge *)coordinator
                                                     textTrigger:(NSString *)textTrigger
 {
     [self removeTriggerTextFromComposer:textTrigger];
-    [self.inputToolbarView pasteText:[UserSuggestionID.room stringByAppendingString:@" "]];
+    [self.inputToolbarView pasteText:[CompletionSuggestionUserID.room stringByAppendingString:@" "]];
+}
+
+- (void)completionSuggestionCoordinatorBridge:(CompletionSuggestionCoordinatorBridge *)coordinator
+                           didRequestCommand:(NSString *)command
+                                 textTrigger:(NSString *)textTrigger
+{
+    [self removeTriggerTextFromComposer:textTrigger];
+    [self setCommand:command];
 }
 
 - (void)removeTriggerTextFromComposer:(NSString *)textTrigger
 {
     RoomInputToolbarView *toolbar = (RoomInputToolbarView *)self.inputToolbarView;
+    Class roomInputToolbarViewClass = [RoomViewController mainToolbarClass];
+
+    // RTE handles removing the text trigger by itself.
+    if (roomInputToolbarViewClass == WysiwygInputToolbarView.class && RiotSettings.shared.enableWysiwygTextFormatting)
+    {
+        return;
+    }
+
     if (toolbar && textTrigger.length) {
         NSMutableAttributedString *attributedTextMessage = [[NSMutableAttributedString alloc] initWithAttributedString:toolbar.attributedTextMessage];
         [[attributedTextMessage mutableString] replaceOccurrencesOfString:textTrigger
@@ -8139,11 +8183,11 @@ static CGSize kThreadListBarButtonItemImageSize;
     }
 }
 
-- (void)userSuggestionCoordinatorBridge:(UserSuggestionCoordinatorBridge *)coordinator didUpdateViewHeight:(CGFloat)height
+- (void)completionSuggestionCoordinatorBridge:(CompletionSuggestionCoordinatorBridge *)coordinator didUpdateViewHeight:(CGFloat)height
 {
-    if (self.userSuggestionContainerHeightConstraint.constant != height)
+    if (self.completionSuggestionContainerHeightConstraint.constant != height)
     {
-        self.userSuggestionContainerHeightConstraint.constant = height;
+        self.completionSuggestionContainerHeightConstraint.constant = height;
 
         [self.view layoutIfNeeded];
     }