+++ Riot/Modules/Room/Settings/RoomSettingsViewController.m
@@ -53,7 +53,6 @@ enum
     ROOM_SETTINGS_MAIN_SECTION_ROW_TAG,
     ROOM_SETTINGS_MAIN_SECTION_ROW_DIRECT_CHAT,
     ROOM_SETTINGS_MAIN_SECTION_ROW_MUTE_NOTIFICATIONS,
-    ROOM_SETTINGS_MAIN_SECTION_ROW_LEAVE,
     //Tchap: Specific rows in room settings
     ROOM_SETTINGS_MAIN_SECTION_ROW_ACCESS_BY_LINK,
     ROOM_SETTINGS_MAIN_SECTION_ROW_ACCESS_RULE
@@ -526,7 +525,6 @@ NSString *const kRoomSettingsAdvancedE2eEnabledCellViewIdentifier = @"kRoomSetti
     }
     [sectionMain addRowWithTag:ROOM_SETTINGS_MAIN_SECTION_ROW_ACCESS_BY_LINK];
     [sectionMain addRowWithTag:ROOM_SETTINGS_MAIN_SECTION_ROW_ACCESS_RULE];
-    [sectionMain addRowWithTag:ROOM_SETTINGS_MAIN_SECTION_ROW_LEAVE];
     [tmpSections addObject:sectionMain];
     
     if (RiotSettings.shared.roomSettingsScreenAllowChangingAccessSettings)
@@ -2412,22 +2410,6 @@ NSString *const kRoomSettingsAdvancedE2eEnabledCellViewIdentifier = @"kRoomSetti
 
                 cell = roomAccessInfo;
             }
-        }
-        else if (row == ROOM_SETTINGS_MAIN_SECTION_ROW_LEAVE)
-        {
-            MXKTableViewCellWithButton *leaveCell = [tableView dequeueReusableCellWithIdentifier:[MXKTableViewCellWithButton defaultReuseIdentifier] forIndexPath:indexPath];
-            
-            NSString* title = [VectorL10n leave];
-            
-            [leaveCell.mxkButton setTitle:title forState:UIControlStateNormal];
-            [leaveCell.mxkButton setTitle:title forState:UIControlStateHighlighted];
-            [leaveCell.mxkButton setTintColor:ThemeService.shared.theme.tintColor];
-            leaveCell.mxkButton.titleLabel.font = [UIFont systemFontOfSize:17];
-            
-            [leaveCell.mxkButton  removeTarget:self action:nil forControlEvents:UIControlEventTouchUpInside];
-            [leaveCell.mxkButton addTarget:self action:@selector(onLeave:) forControlEvents:UIControlEventTouchUpInside];
-            
-            cell = leaveCell;
         }
     }
     else if (section == SECTION_TAG_ACCESS)