+++ Riot/Modules/SecureBackup/Setup/SecureBackupSetupCoordinator.swift
@@ -73,20 +73,22 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
     // MARK: - Public methods
     
     func start() {
-        
-        // Tchap: bypass view offering choice between passphrase/key backup. Tchap only offers key backup option.
-//
-//        let rootViewController = self.createIntro()
-//
-//        if self.navigationRouter.modules.isEmpty == false {
-//            self.navigationRouter.push(rootViewController, animated: true, popCompletion: nil)
-//        } else {
-//            self.navigationRouter.setRootModule(rootViewController)
-//        }        
-        self.showSetupKey(passphraseOnly: false)
-        
+        start(popCompletion: nil)
     }
     
+
+    func start(popCompletion: (() -> Void)?) {
+        // Tchap: bypass view offering choice between passphrase/key backup. Tchap only offers key backup option.
+//        let rootViewController = self.createIntro()
+//        
+//        if self.navigationRouter.modules.isEmpty == false {
+//            self.navigationRouter.push(rootViewController, animated: true, popCompletion: popCompletion)
+//        } else {
+//            self.navigationRouter.setRootModule(rootViewController, popCompletion: popCompletion)
+//        }
+       self.showSetupKey(passphraseOnly: false)
+    }
+
     func toPresentable() -> UIViewController {
         return self.navigationRouter
             .toPresentable()
@@ -105,8 +107,21 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
         return introViewController
     }
     
+    private var dehydrationService: DehydrationService? {
+        if self.session.vc_homeserverConfiguration().encryption.deviceDehydrationEnabled {
+            return self.session.crypto.dehydrationService
+        }
+        
+        return nil
+    }
+    
     private func showSetupKey(passphraseOnly: Bool, passphrase: String? = nil) {
-        let coordinator = SecretsSetupRecoveryKeyCoordinator(recoveryService: self.recoveryService, passphrase: passphrase, passphraseOnly: passphraseOnly, allowOverwrite: allowOverwrite, cancellable: self.cancellable)
+        let coordinator = SecretsSetupRecoveryKeyCoordinator(recoveryService: self.recoveryService,
+                                                             passphrase: passphrase,
+                                                             passphraseOnly: passphraseOnly,
+                                                             allowOverwrite: allowOverwrite,
+                                                             cancellable: self.cancellable,
+                                                             dehydrationService: dehydrationService)
         coordinator.delegate = self
         coordinator.start()